Fix scripting class loading issue #165
Both elasticsearch scripting modules (expression + painless) and
cassandra UDF (see CASSANDRA-11193) need asm jar but it does not seems
possible to only have the asm jar in the cassandra classpath. So, a
JarHell hack was required in order to have the asm jar in both the
cassandra classpath and the elasticsearch scripting modules.